refactor: pull out components from styles files and restructure files
