+++ packages/widget/src/pages/SettingsPage/GasPriceSettings.tsx
@@ -2,11 +2,11 @@ import EvStationIcon from '@mui/icons-material/EvStation';
 import { useTranslation } from 'react-i18next';
 import { useSettings, useSettingsStore } from '../../stores';
 import { Tab, Tabs } from '../../components/Tabs';
-import {
-  BadgedAdditionalInformation,
-  SettingCardExpandable,
-} from './SettingsPage.style';
 import { useSettingMonitor } from '../../hooks';
+import {
+  SettingCardExpandable,
+  BadgedAdditionalInformation,
+} from './SettingsCard';
 
 export const GasPriceSettings: React.FC = () => {
   const { t } = useTranslation();
