+++ packages/widget/src/pages/SettingsPage/SettingsCard/SettingCard.tsx
@@ -1,0 +1,16 @@
+import type { PropsWithChildren } from 'react';
+import { Card } from '../../../components/Card';
+
+export const SettingCard: React.FC<PropsWithChildren> = ({ children }) => {
+  return (
+    <Card
+      variant="default"
+      selectionColor="primary"
+      sx={{
+        p: 2,
+      }}
+    >
+      {children}
+    </Card>
+  );
+};
