+++ packages/widget/src/pages/SettingsPage/SettingsCard/SettingCard.types.ts
@@ -1,0 +1,6 @@
+import type { ReactNode } from 'react';
+
+export interface SettingCardTitle {
+  icon: ReactNode;
+  title: ReactNode;
+}
