+++ packages/widget/src/pages/SettingsPage/SettingsCard/index.ts
@@ -1,0 +1,6 @@
+export * from './SettingCard';
+export * from './SettingCard.style';
+export * from './SettingCardButton';
+export * from './SettingCardContainer';
+export * from './SettingCardExpandable';
+export * from './BadgedAdditionalInformation';
