+++ packages/widget/src/pages/SettingsPage/SlippageSettings/SlippageSettings.style.tsx
@@ -1,9 +1,7 @@
-import type { PropsWithChildren } from 'react';
 import { alpha, styled } from '@mui/material/styles';
-import { Box, ButtonBase, InputBase, Typography } from '@mui/material';
+import { Box, ButtonBase, InputBase } from '@mui/material';
 import type { Theme } from '@mui/material';
 import { inputBaseClasses } from '@mui/material/InputBase';
-import WarningRoundedIcon from '@mui/icons-material/WarningRounded';
 
 export const SettingsFieldSet = styled(Box)(({ theme }) => ({
   display: 'flex',
@@ -90,18 +88,8 @@ export const SlippageCustomInput = styled(InputBase)<SlippageDefaultProps>(({
   };
 });
 
-const SlippageLimitsWarningContainer = styled(Box)(({ theme }) => ({
+export const SlippageLimitsWarningContainer = styled(Box)(({ theme }) => ({
   display: 'flex',
   gap: theme.spacing(1.25),
   marginTop: theme.spacing(1.5),
 }));
-export const SlippageLimitsWarning: React.FC<PropsWithChildren> = ({
-  children,
-}) => (
-  <SlippageLimitsWarningContainer>
-    <WarningRoundedIcon color="warning" />
-    <Typography fontSize={13} fontWeight={400}>
-      {children}
-    </Typography>
-  </SlippageLimitsWarningContainer>
-);
