+++ packages/widget/src/pages/SettingsPage/SlippageSettings/index.ts
@@ -1,0 +1,1 @@
+export * from './SlippageSettings';
